+++ change-log.md
@@ -5,8 +5,9 @@
 ## 2025-06-22
 
 - [x] add modal-transfer component
-- [ ] add options pages (list, detail)
-- [ ] add prompt page-list
+- [x] add options list page
+- [x] add prompt page-list
+- [ ] add options detail page
 - [ ] add prompt page-detail
 - [ ] add info-sample (with remote data)
 - [ ] add desc-sample (with remote data)